+++ src/txn.h
@@ -7,7 +7,7 @@
 typedef void (*txn_callback_t)(void* callback_data, void* response_data);
 
 typedef struct http_request_txn_s {
-  uint16_t txn_id;
+  uint32_t txn_id;
   txn_callback_t callback;
   void* callback_data;
 
@@ -15,7 +15,7 @@ typedef struct http_request_txn_s {
 } txn_t;
 
 typedef struct {
-  uint16_t last_id;
+  uint32_t last_id;
   txn_t* txn_list;
 } txn_cache_t;
 
@@ -27,7 +27,7 @@ txn_t*
 txn_new(txn_cache_t* tc, txn_callback_t callback, void* callback_data);
 
 txn_t*
-txn_find(txn_cache_t* tc, uint16_t txn_id);
+txn_find(txn_cache_t* tc, uint32_t txn_id);
 
 void
 txn_free(txn_cache_t* tc, txn_t* txn);
@@ -36,10 +36,10 @@ void
 txn_callback(txn_t* txn, void* response_data);
 
 void
-txn_update(txn_cache_t* tc, uint16_t txn_id, void* response_data);
+txn_update(txn_cache_t* tc, uint32_t txn_id, void* response_data);
 
 void
-txn_complete(txn_cache_t* tc, uint16_t txn_id, void* response_data);
+txn_complete(txn_cache_t* tc, uint32_t txn_id, void* response_data);
 
 
 #endif
